+++ pkg/storage/storage.go
@@ -18,6 +18,7 @@ import (
 
 var (
 	ErrOverwriteNewerChunk = errors.New("overwriting chunk with newer timestamp")
+	ErrUnknownChunkType    = errors.New("unknown chunk type")
 )
 
 // Result represents the item returned by the read operation, which returns
@@ -292,4 +293,36 @@ func ChunkType(ch swarm.Chunk) swarm.ChunkType {
 		return swarm.ChunkTypeSingleOwner
 	}
 	return swarm.ChunkTypeUnspecified
+}
+
+// IdentityAddress returns the internally used address for the chunk
+// since the single owner chunk address is not a unique identifier for the chunk,
+// but hashing the soc address and the wrapped chunk address is.
+// it is used in the reserve sampling and other places where a key is needed to represent a chunk.
+func IdentityAddress(chunk swarm.Chunk) (swarm.Address, error) {
+
+	if cac.Valid(chunk) {
+		return chunk.Address(), nil
+	}
+
+	// check the chunk is single owner chunk or cac
+	if sch, err := soc.FromChunk(chunk); err == nil {
+		socAddress, err := sch.Address()
+		if err != nil {
+			return swarm.ZeroAddress, err
+		}
+		h := swarm.NewHasher()
+		_, err = h.Write(socAddress.Bytes())
+		if err != nil {
+			return swarm.ZeroAddress, err
+		}
+		_, err = h.Write(sch.WrappedChunk().Address().Bytes())
+		if err != nil {
+			return swarm.ZeroAddress, err
+		}
+
+		return swarm.NewAddress(h.Sum(nil)), nil
+	}
+
+	return swarm.EmptyAddress, fmt.Errorf("identity address failed on chunk %s: %w", chunk, ErrUnknownChunkType)
 }
